In the quest to generate global floodmaps, we've worked on making this go faster. Was able to get average timings from 2.1 sec to 0.8 sec for a 30 meter 1x1 degree DEM tile without bathmetry, 4.3 sec to 2.3 sec with bathymetry.

Changes:

- Numba signatures on some functions help load in much quicker
- pyarrow dependency to speed up CSV/parquet reading
- Improved weightbox creation speed
- Several edits for readability + memory consumption improve performance (such as using np.uint8 instead of np.float arrays)

I have tests to verify that both the burned DEM and the floodmaps generated with this pull request are identical.