@jlgutenson 2025-09-25
Thanks for pointing that out Ricky! I made a few additional changes on top of your and will push those up to the main branch shortly.

I think we can get by with using mainly float32 and int32. If you see something that make you think otherwise, just let me know!